remove duplicate toml::toml namespace
